[ui] Create Avatar component
Moves all the avatar logic and templates into a new
Avatar component.
